# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для взаимодействия с сервисом синхронизации таблиц баз данных.",
        "title": "sync_service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/add-one-conn-bd": {
            "post": {
                "description": "метод для добавления параметров подключения к БД",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "POST"
                ],
                "summary": "AddOneDBConn",
                "parameters": [
                    {
                        "description": "body example",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/urlstorage.JsonFormat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/urlstorage.StorageConnDB"
                        }
                    }
                }
            }
        },
        "/add_sync": {
            "post": {
                "description": "метод для добавления новой синхронизации",
                "tags": [
                    "POST"
                ],
                "summary": "addNewSync",
                "parameters": [
                    {
                        "description": "body example",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.InputDataApi"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/all-connbd": {
            "get": {
                "description": "метод для получения параметров подключения к базам данных и их элиасу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GET"
                ],
                "summary": "getAllDBConn",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/urlstorage.StorageConnDB"
                        }
                    }
                }
            }
        },
        "/all_sync": {
            "get": {
                "description": "some description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GET"
                ],
                "summary": "allSync",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StateAnswer"
                        }
                    }
                }
            }
        },
        "/change-one-conn-bd": {
            "put": {
                "description": "метод для изменения параметров подключения к БД по элиасу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PUT"
                ],
                "summary": "ChangeOneDBConn",
                "parameters": [
                    {
                        "description": "body example",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/urlstorage.JsonFormat"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StateAnswer"
                        }
                    }
                }
            }
        },
        "/one-conn-bd": {
            "post": {
                "description": "метод для получения параметров подключения к БД по элиасу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GET"
                ],
                "summary": "getOneDBConn",
                "parameters": [
                    {
                        "description": "body example",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.RequestDBConn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/urlstorage.ConnDBData"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/start-allSync": {
            "post": {
                "description": "метод для старта всех синхронизаций",
                "tags": [
                    "POST"
                ],
                "summary": "startAllSync",
                "responses": {}
            }
        },
        "/start_sync": {
            "post": {
                "description": "метод для старта приостановленной синхронизации",
                "tags": [
                    "POST"
                ],
                "summary": "startSync",
                "parameters": [
                    {
                        "description": "body example",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.InputDataApi"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/stop-allSync": {
            "post": {
                "description": "метод остановки всех синхронизаций",
                "tags": [
                    "POST"
                ],
                "summary": "stopAllSync",
                "responses": {}
            }
        },
        "/stop_sync": {
            "post": {
                "description": "метод для остановки синхронизации",
                "tags": [
                    "POST"
                ],
                "summary": "stopSync",
                "parameters": [
                    {
                        "description": "body example",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/api.InputDataApi"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
        "api.InputDataApi": {
            "type": "object",
            "properties": {
                "clean": {
                    "type": "boolean"
                },
                "data_base": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "offset": {
                    "type": "string"
                },
                "schema": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "api.RequestDBConn": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                }
            }
        },
        "api.StateAnswer": {
            "type": "object",
            "properties": {
                "data": {},
                "err": {},
                "info": {
                    "type": "string"
                }
            }
        },
        "urlstorage.ConnDBData": {
            "type": "object",
            "properties": {
                "bd_name": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "urlstorage.JsonFormat": {
            "type": "object",
            "properties": {
                "conn_data": {
                    "$ref": "#/definitions/urlstorage.ConnDBData"
                },
                "db_alias": {
                    "type": "string"
                }
            }
        },
        "urlstorage.StorageConnDB": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/urlstorage.ConnDBData"
            }
        }
    }
}